+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # pylint-protobuf
 
+## [0.18.0] - Unreleased
+- Add support for checking missing enum value. This introduces a new warning
+  message "protobuf-enum-value"
+
 ## [0.17.0] - 2021-02-18
 - Add support for scalar repeated fields
 - Fix false-negative warnings on some implicitly defined protobuf message